@@ -1,76 +1,71 @@
 {
   "name": "song-number",
-  "version": "0.1.3",
-  "author": "Florian Chiș",
+  "version": "0.1.4",
+  "author": "Fl0r14n <florian.chis@gmail.com>",
   "homepage": "https://github.com/Fl0r14n/song-number.git",
+  "description": "A small mobile app that will allow user to set a song number from a list of song books and cast it.",
+  "license": "GNUv2",
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:Fl0r14n/song-number.git"
+  },
+  "keywords": [
+    "angular",
+    "song",
+    "google cast",
+    "ionic",
+    "framework"
+  ],
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
+    "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e",
     "electron": "npx cap open @capacitor-community/electron"
   },
   "dependencies": {
-    "@angular/common": "~12.1.1",
-    "@angular/core": "~12.1.1",
-    "@angular/forms": "~12.1.1",
-    "@angular/platform-browser": "~12.1.1",
-    "@angular/platform-browser-dynamic": "~12.1.1",
-    "@angular/router": "~12.1.1",
-    "@capacitor/app": "1.0.2",
-    "@capacitor/camera": "^1.0.3",
-    "@capacitor/core": "3.1.2",
-    "@capacitor/haptics": "1.0.2",
-    "@capacitor/keyboard": "1.0.2",
-    "@capacitor/splash-screen": "^1.1.0",
-    "@capacitor/status-bar": "1.0.2",
-    "@capacitor/storage": "^1.0.3",
-    "@ionic/angular": "^5.5.2",
+    "@angular/common": "~13.0.3",
+    "@angular/core": "~13.0.3",
+    "@angular/forms": "~13.0.3",
+    "@angular/platform-browser": "~13.0.3",
+    "@angular/platform-browser-dynamic": "~13.0.3",
+    "@angular/router": "~13.0.3",
+    "@capacitor/app": "1.0.6",
+    "@capacitor/camera": "^1.2.1",
+    "@capacitor/core": "3.3.2",
+    "@capacitor/haptics": "1.1.3",
+    "@capacitor/keyboard": "1.1.3",
+    "@capacitor/splash-screen": "^1.1.6",
+    "@capacitor/status-bar": "1.0.6",
+    "@capacitor/storage": "^1.2.3",
+    "@ionic/angular": "^5.9.1",
     "@ionic/pwa-elements": "^3.0.2",
-    "@ngx-translate/core": "^13.0.0",
-    "rxjs": "~6.6.3",
-    "tslib": "^2.3.0",
-    "ubuntu-fontface": "^0.1.13",
+    "@ngx-translate/core": "^14.0.0",
+    "rxjs": "~7.4.0",
+    "tslib": "^2.3.1",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.1.1",
-    "@angular-eslint/builder": "~12.0.0",
-    "@angular-eslint/eslint-plugin": "~12.0.0",
-    "@angular-eslint/eslint-plugin-template": "~12.0.0",
-    "@angular-eslint/template-parser": "~12.0.0",
-    "@angular/cli": "~12.1.1",
-    "@angular/compiler": "~12.1.1",
-    "@angular/compiler-cli": "~12.1.1",
-    "@angular/language-service": "~12.0.1",
-    "@capacitor-community/electron": "^3.0.2",
-    "@capacitor/android": "^3.1.2",
-    "@capacitor/cli": "3.1.2",
-    "@capacitor/ios": "^3.1.2",
-    "@ionic/angular-toolkit": "^4.0.0",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "4.16.1",
-    "@typescript-eslint/parser": "4.16.1",
+    "@angular-devkit/build-angular": "~13.0.4",
+    "@angular/cli": "~13.0.4",
+    "@angular/compiler": "~13.0.3",
+    "@angular/compiler-cli": "~13.0.3",
+    "@angular/language-service": "~13.0.3",
+    "@capacitor-community/electron": "^4.0.3",
+    "@capacitor/android": "^3.3.2",
+    "@capacitor/cli": "^3.3.2",
+    "@capacitor/ios": "^3.3.2",
+    "@ionic/angular-toolkit": "^5.0.3",
+    "@types/jasmine": "^3.10.2",
+    "@types/node": "^16.11.11",
     "cordova-chromecast": "git+https://github.com/jellyfin/cordova-plugin-chromecast.git",
-    "eslint": "^7.6.0",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsdoc": "30.7.6",
-    "eslint-plugin-prefer-arrow": "1.2.2",
-    "jasmine-core": "~3.8.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.2",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-coverage-istanbul-reporter": "~3.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "typescript": "~4.2.4"
-  },
-  "description": "A small mobile app that will allow user to set a song number from a list of song books and cast it."
+    "jasmine-core": "^3.10.1",
+    "karma": "^6.3.9",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage": "^2.1.0",
+    "karma-jasmine": "^4.0.1",
+    "karma-jasmine-html-reporter": "^1.7.0",
+    "typescript": "~4.3.5"
+  }
 }
